requirements.txt
pytest
allure-python-commons
Appium-Python-Client
selenium